upgrade to 0.4.4 1. add week_rank_loser updater 2. change trading port settings structure, split trading port into tracking port and backtest port 3. change class Universe structure, add store any used universe in export folder

+++ src/res/factor/fmp/generator/reinforce.py
@@ -116,7 +116,7 @@ class ReinforcePortfolioCreator(PortCreator):
         pool.loc[:, 'rankpct']   = pool['alpha'].rank(pct = True , method = 'first' , ascending = True)
         
         pool = pool.merge(self.init_port.port , on = 'secid' , how = 'left').sort_values('alpha' , ascending = False)
-        pool.loc[:, 'selected'] = pool['weight'].fillna(0) > 0
+        pool.loc[:, 'selected'] = pool['weight'].astype(float).fillna(0) > 0
         pool.loc[:, 'buffered'] = (
             (pool['rankpct'] >= self.conf.buffer_zone) + (pool['selected'].cumsum() <= self.conf.stay_num) *
              (pool['rankpct'] >= self.conf.no_zone))
